pyproject.toml
[tool.poetry]
name = "cypher-cloud"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.115.6"
uvicorn = {extras = ["standard"], version = "^0.32.1"}
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
pyotp = "^2.9.0"
cryptography = "^44.0.0"
python-multipart = "^0.0.19"
aiofiles = "^24.1.0"
sqlalchemy = {extras = ["asyncio"], version = "^2.0.36"}
psycopg2-binary = "^2.9.10"
pydantic = {extras = ["email"], version = "^2.10.3"}
asyncpg = "^0.30.0"
pydantic-settings = "^2.6.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
